+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Note: the BigBuy API is not versioned and some endpoints are often added or removed without notice.
 The removal of methods that call deleted endpoints is not considered a breaking change.
 
-## Unreleased
+## 3.19.2 (2023/10/19)
 
 * Add `get_user_auth_status`, `get_product_stock_by_handling_days`, `get_product_variations_stock_by_handling_days`
 * Deprecate `get_products_stock`, `get_product_stock`, `get_products_variations_stock`